Fix bug for crontab feature when user's cron table is empty (#690)
## DESCRIPTION OF CHANGES: 
When a user's `cron` table is empty, the `get_crontab_contents` function does not set the return variable properly due to a bug in the way `printf -v ...` is called towards the end of that function.  This PR fixes that bug as well as some others found after testing with an initially empty user cron table.

## TESTS CONDUCTED: 
On Hera, Jet, and Cheyenne, ran the WE2E test `deactivate_tasks`, starting with both an empty `cron` table and a non-empty one.  All tests passed with the `cron` table being modified properly at the start and end of the test.

## CONTRIBUTORS (optional): 
@chan-hoo pointed out this error on Hera.