Bug fixes in machine files (#692)
## DESCRIPTION OF CHANGES: 
Cleaning up bugs in the machine files.  The first bug prompted this PR, and the rest were found subsequently.  The bugs (and their fixes) are as follows:

1) A space is missing after the `print_info_msg` and `print_err_msg_exit` function calls in the `file_location` functions.  Inserting a space gets passed this bug, but subsequent issues were found as described below.

**For machine files that call the `print_info_msg` function in `file_location` (`cheyenne.sh`, `hera.sh`, `jet.sh`, and `orion.sh`):**
Fixing this bug leads to other failures because when the "*" stanza is encountered in the `file_location` function, 
the `EXTRN_MDL_SYSBASEDIR_ICS|LBCS` variable gets set to the message that `file_location` returns.  Since that message contains spaces, it leads to other failures in downstream scripts (the ex-scripts).  Simply removing the printing out of the message (thus causing `EXTRN_MDL_SYSBASEDIR_ICS|LBCS` to be set to a null string) fixes the failures, so this was the fix implemented.  If desired, a message for an empty value for `EXTRN_MDL_SYSBASEDIR_ICS|LBCS` can be placed in another script (where those variables are used).

**For machine files that use `print_err_msg_exit` in `file_location` (`stampede.sh` and `wcoss_dell_p3.sh`):**
These should not exit if the file location is not available since the experiment can still complete successfully.  So just removing the `print_err_msg_exit` call should work (and make the behavior of these machine files consistent with the set above).

2) In all the machine files, the variable `FV3GFS_FILE_FMT_ICS` should be changed to `FV3GFS_FILE_FMT_LBCS` in the definition of `EXTRN_MDL_SYSBASEDIR_LBCS`.  This was fixed in all the files.

3) In `stampede.sh`, a variable named `SYSBASEDIR_ICS` is defined.  This is a typo.  Modify to `EXTRN_MDL_SYSBASEDIR_ICS`.

## TESTS CONDUCTED: 
Ran the WE2E test `grid_RRFS_CONUS_25km_ics_HRRR_lbcs_RAP_suite_GSD_SAR` on:
* Hera -- successful
* Jet -- successful except for UPP tasks
* Cheyenne -- successful except for UPP tasks

The UPP task failures are new and being experienced by other PRs as well (e.g. #689).  The original issue with machine files seems resolved.

## CONTRIBUTORS (optional): 
@JeffBeck-NOAA encountered and reported the original error.